refactor(ios): extract browser auth plumbing from AuthScreen (#14272)
## Summary

Iteration 9 of the iOS refactor loop (no Linear issue — ad-hoc loop work). Behavior-preserving split of `Features/Auth/AuthScreen.swift` (**517 → 232 lines**).

- The non-view browser-auth machinery moves to a new `Features/Auth/MobileAuthCoordinator.swift` (286 lines): `MobileBrowserAuthURLBuilder` (PKCE verifier/challenge generation), `MobileAuthCoordinatorError`, `MobileAuthCoordinator` (ASWebAuthenticationSession presentation), `isAuthSessionCancellation`, and the private `String`/`Data` base64url helpers.
- Access changes: only `isAuthSessionCancellation` goes `private` → internal (called from the view). All other bodies verbatim.
- `AuthScreen.swift` keeps the view + private subviews; drops now-unused `AuthenticationServices`/`CryptoKit`/`Security`/`UIKit` imports.
- `project.pbxproj` wired with explicit entries (`plutil -lint` clean).

## Verification

- `pnpm run ios:lint` → **clean ✓**
- `pnpm run ios:test` (iPhone 17 sim, iOS 26.3.1) → **248 passed, 6 skipped, 2 failed** — the known local-sim-only pair tracked in **JOV-4252** (passes on GitHub CI).
- `AuthScreenStyleGuardTests` and `MobileAuthFinalizationTests` pass.

## Risks

- Low: same-module move of auth plumbing, one access widening, no logic changes. Auth-adjacent by file location only.
